EDU-583 Unify urls and paths utils modules
