+++ src/plugins/pdf-viewer/pdf-viewer-info.js
@@ -3,7 +3,7 @@ import { SOURCE_TYPE } from './constants.js';
 import PdfViewerIcon from './pdf-viewer-icon.js';
 import cloneDeep from '../../utils/clone-deep.js';
 import PdfViewerDisplay from './pdf-viewer-display.js';
-import { isAccessibleStoragePath } from '../../ui/path-helper.js';
+import { isAccessibleStoragePath } from '../../utils/storage-utils.js';
 
 class PdfViewerInfo {
   static get typeName() { return 'pdf-viewer'; }
