--- src/utils/path-utils.js
@@ -1,14 +1,0 @@
-import path from 'path';
-import urlUtils from './url-utils.js';
-import uniqueId from './unique-id.js';
-import slugify from '@sindresorhus/slugify';
-
-export const componseUniqueFileName = (fileName, parentPath = null) => {
-  const id = uniqueId.create();
-  const extension = path.extname(fileName);
-  const baseName = fileName.substr(0, fileName.length - extension.length);
-  const slugifiedBaseName = slugify(baseName);
-  const uniqueBaseName = [slugifiedBaseName, id].filter(x => x).join('-');
-  const newFileName = `${uniqueBaseName}${extension}`;
-  return parentPath ? urlUtils.concatParts(parentPath, newFileName) : newFileName;
-};
